Finish up the game (mostly)
I basically finished the game.
Added a while loop in game() to play additional rounds in case of a draw,
so that the final result would be either a win or a loss, no ties.

I also removed the option for the player to choose the number of rounds
as it seemed unnecessary and it provided the option for the player to
enter an even number of rounds.

There are a couple of problems I need to address next - code a behavior
for when the player clicks the cancel button/presses the esc key during
the game (it now returns a type error because of the .toLowerCase()
method); and also to account for the cases when the player accidentally
makes a typo (right now any input different from rock, paper, scissors
is considered a win for the computer).