Finish modal
I made some changes to the styling of the modal to make it fit better
with the overal looks of the page.

I added an event listener so that the player can close the modal by clicking
outside its borders

This made me create a couple of functions to set and remove the
modal-active class instead of simply using the toggle method as I intended
at first.
I opted for this somewhat more "redundant" solution because using the
toggle method would have resulted in some unexpected behaviour
when the player closes the modal by clicking outside the box and then
clicks the reset button during the game - it would have made the modal
visible upon reset, when it should only pop up at game over.